Detect anon users, report custom fields
Several issues were preventing proper import of data into jira,
including the fact that numerous users don't exist in Jira. The
anonymous user handling routines follow the strategy employed by the
Jira mail handler when createusers=false and reporterusername is set

Comments and descriptions are prepended with the anon user's email
address if it's available

Jira also freaks if you use a value that isn't defined on multi select
fields, so we print those out at the end along.

Users that are treated as anonymous are printed as well